refactor: 토너먼트 진행 가이드 팝오버 분리
